Add files via upload
New spreadsheet for R code. The following changes have been made to the pre-existing spreadsheet:

Column A: Addition of column “Project #” and reordering of all other columns accordingly.

Columns G-O: Addition of columns for secondary analysis on engagement methods.

Columns Q-U: Addition of columns for secondary analysis on social media platforms. 

Column X: Correction of typo “Rgular” to “Regular”

Columns AB-AE: Addition of columns for secondary analysis on training methods (unused in analysis).

Column AL: Correction of typo “51-100” to “51 – 100”

Column AM: “1 – 5” changed to “1 to 5”, “6 – 10” changed to “6 to 10” and “11 – 20” changed to “11 to 20” to force a stop to Excel’s transformation of ranges to dates.